pyproject.toml
[tool.poetry]
name = "scan-print-pipeline"
version = "0.1.0"
description = ""
authors = ["Christopher Carroll Smith <75859865+chriscarrollsmith@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "scan_print_pipeline"}]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.4.0"
openai = "^0.27.8"
epson-connect = "^0.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
